Rename view sequencer test module to be consistent with library modules
